+++ client/src/Componentes/Home/Home.module.css
@@ -1,0 +1,9 @@
+.home {
+  display: flex;
+  flex-direction: column;
+  height: 100vh;
+}
+
+.footer {
+  margin-top: auto;
+}
