+++ client/src/Componentes/Pagination/Pagination.module.css
@@ -5,6 +5,7 @@
   position: relative;
   bottom: 0;
   width: 100%;
+  margin-bottom: 50px;
 }
 
 .paginationUl {
